requirements.txt
discord==2.1.0
openai==0.25.0
python-dotenv==0.21.0
tenacity==8.0.1
